layer2/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
scikit-learn>=1.5.0
numpy>=1.26.0
pydantic>=2.0.0
